Move remaining rebalancing tests to PartitionBalancerTest
